requirements.txt
requests~=2.31.0
pandas~=3.0.2
anthropic~=0.96.0
openpyxl
python-dotenv
